Address endpoints completed
